Fixed files

+++ src/main/java/com/example/Proect/Repositories/EnemyPokemonRepository.java
@@ -3,5 +3,5 @@ package com.example.Proect.Repositories;
 import com.example.Proect.Entities.EnemyPokemon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnemyPokemonRepository extends JpaRepository<EnemyPokemon, Integer> {
+public interface EnemyPokemonRepository extends JpaRepository<EnemyPokemon, Long> {
 }
